GH_ISSUE_902: split objects_write.go by operation type
The single 893-line file mixed PUT, COPY (native + materialized +
ambiguous-recovery), DELETE single + batch, materialization helpers,
and finalize helpers. Local reasoning + review were both getting
harder per #902.

Splits the file into five focused siblings (issue-suggested layout):

  - put.go: PutObject, attemptPutOnBackend, bufferPutBody,
    buildPutPayload, withoutBackend, putAttemptRequest/Result,
    errDrainRaceAborted.
  - copy.go: CopyObject, headSourceForCopy, sameBackendCopyEligible,
    tryNativeCopy, probeDestAfterAmbiguousCopy, nativeCopyRequest.
  - delete.go: DeleteObject, DeleteObjects, flattenBatchDeletes,
    tallyDeleteResults, defaultBatchDeleteConcurrency,
    DeleteObjectResult, batchDeleteItem.
  - mutation_finalize.go: finalizePutSuccess, finalizeMaterialized-
    Copy, finalizeNativeCopy, finalizeDelete, finalizeBatchDelete and
    the two request structs that exist solely to bundle finalize args
    (putSuccessRequest, materializedCopyRequest).
  - materialize.go (existing): materializedSource +
    materializeCopySource + tryMaterializeFromLocation appended here
    since they share file-purpose with materializedBody.

Pure structural move - no behaviour change, no test changes.
