+++ .version
@@ -1,1 +1,1 @@
-v0.60.4
+v0.60.5
